+++ changelog.md
@@ -15,6 +15,7 @@
 - apps menu and task switcher not working when used on fullscreen apps.
 - dock touch support.
 - toolbar and dock showing while fullscreen on multimonitor setups.
+- twm pause not working as expected.
 
 ## [2.6.8]